+++ changelog.markdown
@@ -147,3 +147,4 @@
 * Simplified file **src-test/Main.hs**.
 * Added file **.travis.yml** to control testing by <https://docs.travis-ci.com>.
 * Added file **.ghci**.
+* Replaced use of module **ToolShed.Defaultable** with **Data.Default**.